seperated edit post moda, able to edit tags properly
